backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click-spinner==0.1.10
click8==8.0.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.1.7
django-celery-results==2.5.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.0
phonenumbers==8.13.9
Pillow==9.4.0
pindo-cli==0.2.1
prompt-toolkit==3.0.38
pycparser==2.21
Pygments==2.15.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rich==10.16.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomlkit==0.7.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
